Traffic-Sign-Recognition-using-convolution-neural-network-with-RTMaps-Embedded-and-BlueBox-V2
The main goal of this project is to establish a neural network in RTMaps-embedded python block to identify the traffic sign board from the input images given. This project uses BlueBox 2.0 and RTMaps-Embedded as a software. The communication between RTMaps and Bluebox is set with TCP IP Connection. The main scope of this project is to identify the given input traffic sign images and print a label of the image under which category it comes under. We have categorized the images into 43 classes. Firstly, a convolution neural network is established and trained with the German traffic sign data set for about ~39000 images. Once the neural network is trained, it is then added to the RTMaps-Embedded python block. Along with that we also give the input images to be identified. So, finally the model created as RTMaps Component classifies the input traffic sign given to one of the 43 classes.